refactor: mnist train
