--- train/download_pretrained_model.py
@@ -1,32 +1,0 @@
-import os
-import urllib.request
-from urllib.error import HTTPError
-
-# Github URL where saved models are stored for this tutorial
-base_url = "https://raw.githubusercontent.com/phlippe/saved_models/main/tutorial5/"
-# Files to download (only ResNet model files)
-pretrained_files = [
-    "ResNet.ckpt",
-    "tensorboards/ResNet/events.out.tfevents.resnet",
-]
-# Path to save the checkpoints
-CHECKPOINT_PATH = "./checkpoints"
-
-# Create checkpoint path if it doesn't exist yet
-os.makedirs(CHECKPOINT_PATH, exist_ok=True)
-
-# For each file, check whether it already exists. If not, try downloading it.
-for file_name in pretrained_files:
-    file_path = os.path.join(CHECKPOINT_PATH, file_name)
-    if "/" in file_name:
-        os.makedirs(file_path.rsplit("/", 1)[0], exist_ok=True)
-    if not os.path.isfile(file_path):
-        file_url = base_url + file_name
-        print(f"Downloading {file_url}...")
-        try:
-            urllib.request.urlretrieve(file_url, file_path)
-        except HTTPError as e:
-            print(
-                "Something went wrong. Please try to download the file from the GDrive folder, or contact the author with the full output including the following error:\n",
-                e,
-            )
