--- train/train.py
@@ -1,313 +1,0 @@
-import os
-from types import SimpleNamespace
-
-import pytorch_lightning as pl
-import torch
-import torch.nn as nn
-import torch.optim as optim
-import torch.utils.data as data
-from torchvision import transforms
-from torchvision.datasets import MNIST
-
-DATASET_PATH = os.environ.get("PATH_DATASETS", "data/")
-CHECKPOINT_PATH = os.environ.get("PATH_CHECKPOINT", "checkpoints")
-
-pl.seed_everything(42)
-
-torch.backends.cudnn.deterministic = True
-torch.backends.cudnn.benchmark = False
-torch.device("cpu")
-
-train_dataset = MNIST(root=DATASET_PATH, train=True, download=False)
-DATA_MEANS = (train_dataset.data / 255.0).mean(axis=(0, 1, 2))
-DATA_STD = (train_dataset.data / 255.0).std(axis=(0, 1, 2))
-
-test_transform = transforms.Compose(
-    [transforms.ToTensor(), transforms.Normalize(DATA_MEANS, DATA_STD)]
-)
-train_transform = transforms.Compose(
-    [
-        transforms.RandomHorizontalFlip(),
-        transforms.RandomResizedCrop((32, 32), scale=(0.8, 1.0), ratio=(0.9, 1.1)),
-        transforms.ToTensor(),
-        transforms.Normalize(DATA_MEANS, DATA_STD),
-    ]
-)
-train_dataset = MNIST(
-    root=DATASET_PATH, train=True, transform=train_transform, download=False
-)
-val_dataset = MNIST(
-    root=DATASET_PATH, train=True, transform=test_transform, download=False
-)
-pl.seed_everything(42)
-train_set, _ = torch.utils.data.random_split(train_dataset, [54000, 6000])
-pl.seed_everything(42)
-_, val_set = torch.utils.data.random_split(val_dataset, [54000, 6000])
-
-test_set = MNIST(
-    root=DATASET_PATH, train=False, transform=test_transform, download=False
-)
-
-train_loader = data.DataLoader(
-    train_set,
-    batch_size=128,
-    shuffle=True,
-    drop_last=True,
-    pin_memory=True,
-    num_workers=0,
-)
-val_loader = data.DataLoader(
-    val_set, batch_size=128, shuffle=False, drop_last=False, num_workers=0
-)
-test_loader = data.DataLoader(
-    test_set, batch_size=128, shuffle=False, drop_last=False, num_workers=0
-)
-
-
-class MNISTModule(pl.LightningModule):
-    def __init__(self, model_name, model_hparams, optimizer_name, optimizer_hparams):
-        """MNISTModule.
-
-        Args:
-            model_name: Name of the model/CNN to run. Used for creating the model (see function below)
-            model_hparams: Hyperparameters for the model, as dictionary.
-            optimizer_name: Name of the optimizer to use. Currently supported: Adam, SGD
-            optimizer_hparams: Hyperparameters for the optimizer, as dictionary. This includes learning rate, weight decay, etc.
-
-        """
-        super().__init__()
-        self.save_hyperparameters()
-        self.model = create_model(model_name, model_hparams)
-        self.loss_module = nn.CrossEntropyLoss()
-        self.example_input_array = torch.zeros((1, 1, 28, 28), dtype=torch.float32)
-
-    def forward(self, imgs):
-        return self.model(imgs)
-
-    def configure_optimizers(self):
-        if self.hparams.optimizer_name == "Adam":
-            optimizer = optim.AdamW(self.parameters(), **self.hparams.optimizer_hparams)
-        elif self.hparams.optimizer_name == "SGD":
-            optimizer = optim.SGD(self.parameters(), **self.hparams.optimizer_hparams)
-        else:
-            raise ValueError(f'Unknown optimizer: "{self.hparams.optimizer_name}"')
-
-        scheduler = optim.lr_scheduler.MultiStepLR(
-            optimizer, milestones=[100, 150], gamma=0.1
-        )
-        return [optimizer], [scheduler]
-
-    def training_step(self, batch, batch_idx):
-        imgs, labels = batch
-        preds = self.model(imgs)
-        loss = self.loss_module(preds, labels)
-        acc = (preds.argmax(dim=-1) == labels).float().mean()
-        self.log("train_acc", acc, on_step=False, on_epoch=True)
-        self.log("train_loss", loss)
-        return loss
-
-    def validation_step(self, batch, batch_idx):
-        imgs, labels = batch
-        preds = self.model(imgs).argmax(dim=-1)
-        acc = (labels == preds).float().mean()
-        self.log("val_acc", acc)
-
-    def test_step(self, batch, batch_idx):
-        imgs, labels = batch
-        preds = self.model(imgs).argmax(dim=-1)
-        acc = (labels == preds).float().mean()
-        self.log("test_acc", acc)
-
-
-from pytorch_lightning.callbacks import LearningRateMonitor, ModelCheckpoint
-
-model_dict = {}
-
-
-def create_model(model_name, model_hparams):
-    if model_name in model_dict:
-        return model_dict[model_name](**model_hparams)
-    else:
-        raise ValueError(f'Unknown model name "{model_name}".')
-
-
-act_fn_by_name = {
-    "tanh": nn.Tanh,
-    "relu": nn.ReLU,
-    "leakyrelu": nn.LeakyReLU,
-    "gelu": nn.GELU,
-}
-
-
-def train_model(model_name, save_name=None, **kwargs):
-    """Train model.
-
-    Args:
-        model_name: Name of the model you want to run. Is used to look up the class in "model_dict"
-        save_name (optional): If specified, this name will be used for creating the checkpoint and logging directory.
-
-    """
-    if save_name is None:
-        save_name = model_name
-
-    trainer = pl.Trainer(
-        default_root_dir=os.path.join(CHECKPOINT_PATH, save_name),
-        accelerator="auto",
-        devices=1,
-        max_epochs=10,
-        callbacks=[
-            ModelCheckpoint(save_weights_only=True, mode="max", monitor="val_acc"),
-            LearningRateMonitor("epoch"),
-        ],
-    )
-    trainer.logger._log_graph = True
-    trainer.logger._default_hp_metric = None
-
-    pretrained_filename = os.path.join(CHECKPOINT_PATH, save_name + ".ckpt")
-    if os.path.isfile(pretrained_filename):
-        print(f"Found pretrained model at {pretrained_filename}, loading...")
-        model = MNISTModule.load_from_checkpoint(pretrained_filename)
-    else:
-        pl.seed_everything(42)
-        model = MNISTModule(model_name=model_name, **kwargs)
-        trainer.fit(model, train_loader, val_loader)
-        model = MNISTModule.load_from_checkpoint(
-            trainer.checkpoint_callback.best_model_path
-        )
-
-    val_result = trainer.test(model, dataloaders=val_loader, verbose=False)
-    test_result = trainer.test(model, dataloaders=test_loader, verbose=False)
-    result = {"test": test_result[0]["test_acc"], "val": val_result[0]["test_acc"]}
-
-    return model, result
-
-
-class ResNetBlock(nn.Module):
-    def __init__(self, c_in, act_fn, subsample=False, c_out=-1):
-        """ResNetBlock.
-
-        Args:
-            c_in: Number of input features
-            act_fn: Activation class constructor (e.g. nn.ReLU)
-            subsample - If True, we want to apply a stride inside the block and reduce the output shape by 2 in height and width
-            c_out - Number of output features. Note that this is only relevant if subsample is True, as otherwise, c_out = c_in
-
-        """
-        super().__init__()
-        if not subsample:
-            c_out = c_in
-        self.net = nn.Sequential(
-            nn.Conv2d(
-                c_in,
-                c_out,
-                kernel_size=3,
-                padding=1,
-                stride=1 if not subsample else 2,
-                bias=False,
-            ),
-            nn.BatchNorm2d(c_out),
-            act_fn(),
-            nn.Conv2d(c_out, c_out, kernel_size=3, padding=1, bias=False),
-            nn.BatchNorm2d(c_out),
-        )
-        self.downsample = (
-            nn.Conv2d(c_in, c_out, kernel_size=1, stride=2) if subsample else None
-        )
-        self.act_fn = act_fn()
-
-    def forward(self, x):
-        z = self.net(x)
-        if self.downsample is not None:
-            x = self.downsample(x)
-        return self.act_fn(z + x)
-
-
-resnet_blocks_by_name = {"ResNetBlock": ResNetBlock}
-
-
-class ResNet(nn.Module):
-    def __init__(
-        self,
-        num_classes=10,
-        num_blocks=[3, 3, 3],
-        c_hidden=[16, 32, 64],
-        act_fn_name="relu",
-        block_name="ResNetBlock",
-    ):
-        """ResNet.
-
-        Args:
-            num_classes - Number of classification outputs (10 for MNIST)
-            num_blocks - List with the number of ResNet blocks to use. The first block of each group uses downsampling, except the first.
-            c_hidden - List with the hidden dimensionalities in the different blocks. Usually multiplied by 2 the deeper we go.
-            act_fn_name - Name of the activation function to use, looked up in "act_fn_by_name"
-            block_name - Name of the ResNet block, looked up in "resnet_blocks_by_name"
-
-        """
-        super().__init__()
-        self.hparams = SimpleNamespace(
-            num_classes=num_classes,
-            c_hidden=c_hidden,
-            num_blocks=num_blocks,
-            act_fn_name=act_fn_name,
-            act_fn=act_fn_by_name[act_fn_name],
-            block_class=resnet_blocks_by_name[block_name],
-        )
-        self._create_network()
-        self._init_params()
-
-    def _create_network(self):
-        c_hidden = self.hparams.c_hidden
-        self.input_net = nn.Sequential(
-            nn.Conv2d(1, c_hidden[0], kernel_size=3, padding=1, bias=False),
-            nn.BatchNorm2d(c_hidden[0]),
-            self.hparams.act_fn(),
-        )
-        blocks = []
-        for block_idx, block_count in enumerate(self.hparams.num_blocks):
-            for bc in range(block_count):
-                subsample = bc == 0 and block_idx > 0
-                blocks.append(
-                    self.hparams.block_class(
-                        c_in=c_hidden[block_idx if not subsample else (block_idx - 1)],
-                        act_fn=self.hparams.act_fn,
-                        subsample=subsample,
-                        c_out=c_hidden[block_idx],
-                    )
-                )
-        self.blocks = nn.Sequential(*blocks)
-        self.output_net = nn.Sequential(
-            nn.AdaptiveAvgPool2d((1, 1)),
-            nn.Flatten(),
-            nn.Linear(c_hidden[-1], self.hparams.num_classes),
-        )
-
-    def _init_params(self):
-        for m in self.modules():
-            if isinstance(m, nn.Conv2d):
-                nn.init.kaiming_normal_(
-                    m.weight, mode="fan_out", nonlinearity=self.hparams.act_fn_name
-                )
-            elif isinstance(m, nn.BatchNorm2d):
-                nn.init.constant_(m.weight, 1)
-                nn.init.constant_(m.bias, 0)
-
-    def forward(self, x):
-        x = self.input_net(x)
-        x = self.blocks(x)
-        return self.output_net(x)
-
-
-model_dict["ResNet"] = ResNet
-
-resnet_model, resnet_results = train_model(
-    model_name="ResNet",
-    model_hparams={
-        "num_classes": 10,
-        "c_hidden": [16, 32, 64],
-        "num_blocks": [3, 3, 3],
-        "act_fn_name": "relu",
-    },
-    optimizer_name="SGD",
-    optimizer_hparams={"lr": 0.1, "momentum": 0.9, "weight_decay": 1e-4},
-)
