+++ train_mnist/src/configs.py
@@ -1,0 +1,21 @@
+import os
+
+
+class Configs:
+    DATASET_PATH = os.environ.get(
+        "PATH_DATASETS",
+        "/Users/aimmo-aiy-0297/Desktop/workspace/mlops/train_mnist/data",
+    )
+    CHECKPOINT_PATH = os.environ.get(
+        "PATH_CHECKPOINT",
+        "/Users/aimmo-aiy-0297/Desktop/workspace/mlops/train_mnist/checkpoints",
+    )
+    RANDOM_SEED = os.environ.get("RANDOM_SEED", 10)
+    OPTIMIZER = os.environ.get("OPRIMIZER", "SGD")
+    LEARNING_RATE = os.environ.get("LEARNING_RATE", 0.1)
+    MOMENTUM = os.environ.get("MOMENTUM", 0.9)
+    WEIGHT_DECAY = os.environ.get("WEIGHT_DECAY", 1e-4)
+    ACT_FN_NAME = os.environ.get("ACT_FN_NAME", "relu")
+    C_HIDDEN = os.environ.get("C_HIDDEN", [16, 32, 64])
+    NUM_BLOCKS = os.environ.get("NUM_BLOCKS", [3, 3, 3])
+    EPOCHS = os.environ.get("EPOCHS", 1)
