+++ train_mnist/src/models/__init__.py
@@ -1,0 +1,3 @@
+from models.resnet import ResNet
+
+MODEL_DICT = {"ResNet": ResNet}
