+++ train_mnist/src/models/resnet.py
@@ -1,0 +1,126 @@
+from types import SimpleNamespace
+
+import torch.nn as nn
+
+
+class ResNetBlock(nn.Module):
+    def __init__(self, c_in, act_fn, subsample=False, c_out=-1):
+        """ResNetBlock.
+
+        Args:
+            c_in: Number of input features
+            act_fn: Activation class constructor (e.g. nn.ReLU)
+            subsample - If True, we want to apply a stride inside the block and reduce the output shape by 2 in height and width
+            c_out - Number of output features. Note that this is only relevant if subsample is True, as otherwise, c_out = c_in
+
+        """
+        super().__init__()
+        if not subsample:
+            c_out = c_in
+        self.net = nn.Sequential(
+            nn.Conv2d(
+                c_in,
+                c_out,
+                kernel_size=3,
+                padding=1,
+                stride=1 if not subsample else 2,
+                bias=False,
+            ),
+            nn.BatchNorm2d(c_out),
+            act_fn(),
+            nn.Conv2d(c_out, c_out, kernel_size=3, padding=1, bias=False),
+            nn.BatchNorm2d(c_out),
+        )
+        self.downsample = (
+            nn.Conv2d(c_in, c_out, kernel_size=1, stride=2) if subsample else None
+        )
+        self.act_fn = act_fn()
+
+    def forward(self, x):
+        z = self.net(x)
+        if self.downsample is not None:
+            x = self.downsample(x)
+        return self.act_fn(z + x)
+
+
+resnet_blocks_by_name = {"ResNetBlock": ResNetBlock}
+
+act_fn_by_name = {
+    "tanh": nn.Tanh,
+    "relu": nn.ReLU,
+    "leakyrelu": nn.LeakyReLU,
+    "gelu": nn.GELU,
+}
+
+
+class ResNet(nn.Module):
+    def __init__(
+        self,
+        num_classes=10,
+        num_blocks=[3, 3, 3],
+        c_hidden=[16, 32, 64],
+        act_fn_name="relu",
+        block_name="ResNetBlock",
+    ):
+        """ResNet.
+
+        Args:
+            num_classes - Number of classification outputs (10 for MNIST)
+            num_blocks - List with the number of ResNet blocks to use. The first block of each group uses downsampling, except the first.
+            c_hidden - List with the hidden dimensionalities in the different blocks. Usually multiplied by 2 the deeper we go.
+            act_fn_name - Name of the activation function to use, looked up in "act_fn_by_name"
+            block_name - Name of the ResNet block, looked up in "resnet_blocks_by_name"
+
+        """
+        super().__init__()
+        self.hparams = SimpleNamespace(
+            num_classes=num_classes,
+            c_hidden=c_hidden,
+            num_blocks=num_blocks,
+            act_fn_name=act_fn_name,
+            act_fn=act_fn_by_name[act_fn_name],
+            block_class=resnet_blocks_by_name[block_name],
+        )
+        self._create_network()
+        self._init_params()
+
+    def _create_network(self):
+        c_hidden = self.hparams.c_hidden
+        self.input_net = nn.Sequential(
+            nn.Conv2d(1, c_hidden[0], kernel_size=3, padding=1, bias=False),
+            nn.BatchNorm2d(c_hidden[0]),
+            self.hparams.act_fn(),
+        )
+        blocks = []
+        for block_idx, block_count in enumerate(self.hparams.num_blocks):
+            for bc in range(block_count):
+                subsample = bc == 0 and block_idx > 0
+                blocks.append(
+                    self.hparams.block_class(
+                        c_in=c_hidden[block_idx if not subsample else (block_idx - 1)],
+                        act_fn=self.hparams.act_fn,
+                        subsample=subsample,
+                        c_out=c_hidden[block_idx],
+                    )
+                )
+        self.blocks = nn.Sequential(*blocks)
+        self.output_net = nn.Sequential(
+            nn.AdaptiveAvgPool2d((1, 1)),
+            nn.Flatten(),
+            nn.Linear(c_hidden[-1], self.hparams.num_classes),
+        )
+
+    def _init_params(self):
+        for m in self.modules():
+            if isinstance(m, nn.Conv2d):
+                nn.init.kaiming_normal_(
+                    m.weight, mode="fan_out", nonlinearity=self.hparams.act_fn_name
+                )
+            elif isinstance(m, nn.BatchNorm2d):
+                nn.init.constant_(m.weight, 1)
+                nn.init.constant_(m.bias, 0)
+
+    def forward(self, x):
+        x = self.input_net(x)
+        x = self.blocks(x)
+        return self.output_net(x)
